extract write_batch
